// src/app/core/Enums/yetkiGrupEnum.ts
export enum YetkiGrupEnum {
    menu = "Menüler",
    menuRecourse = "Menü Başvurular",
    menuUnitManagement = "Menü Birim Yönetimi",
    menuStudent = "Menü Student",
    menuGeneralManagement = "Menü Genel Yönetim",
    dashboard = "Dashboard",

    banka = "Banka",
    birim = "Birim",
    dil = "Dil",//*
    konsolosluk = "Konsolosluk",
    kullanici = "Kullanıcı",
    kurum = "Kurum",
    ulke = "Ülke",
    yetki = "Yetki",
    settings = "Settings",

    ogrenci = "Öğrenci",
    ogrenciSinavBilgisi="Öğrenci Sınav Bilgisi",

    birimBolumleri = "B.Y. Bölüm",//*
    birimDilSinavlari = "B.Y. Dil Sınavları",//*
    birimEgitimDonemleri = "B.Y. Eğitim Dönemleri",//*
    birimFakulteleri = "B.Y. Fakülteler",//*
    birimDilleri = "B.Y. Dilleri",
    birimKabulKosullari = "B.Y. Kabul Koşulları",
    birimPuanTurleri = "B.Y. Puan Türleri",
    birimSablon = "B.Y. Şablonları",
    
    bSinav = "Başvuru Sınav",//başvuru sınav *
    bOrtak = "Başvuru Ortak",//başvuru ortak*
    bLisans = "Başvuru Lisans",
    bYuksekLisans = "Başvuru Yüksek Lisans",
    bDoktora = "Başvuru Doktora",
    bGirisUcreti = "Başvuru Giriş Ücreti",
    bBildirim= "Başvuru Bildirim",

    bSinavMerkezi = "Başvuru Sınav Merkezi",
    bSinavMerkeziSalonu="Başvuru Sınav Merkezi Salonları",

    bSinavDigerIslem = "Başvuru Sınav İşlemleri",
    bOrtakDigerIslem = "Başvuru Ortak İşlemler",

    bOgrenciler="Başvuru Öğrencileri",
    bSinavOgrenciler="Sınav Öğrencileri",

    bTercihBirimBolum="Başvuru Tercih Birim Bölümler",
    bTercihBBKontenjan="Başvuru Tercih Birim Bölüm Kontenjanlar", 

}